Break out shard logic from editor and logs component
